move test harnesses to testing dir
