+++ packages/toolbar/src/utils/index.ts
@@ -3,4 +3,3 @@ export {
   getLgIds,
   type GetLgIdsReturnType,
 } from './getLgIds';
-export { getTestUtils, type TestUtilsReturnType } from './getTestUtils';
